Mailer: Move Builder classes to Builder folder #34
