+++ src/Service/Mailer/Builder/MailerBuilder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service\Mailer;
+namespace App\Service\Mailer\Builder;
 
 use Swift_Mailer;
